Add subscribeToChangedState updates extension fun.
This subscribe variant will receive only distinctUntilChanged state updates.
