+++ library/src/main/kotlin/com/freeletics/coredux/ReduxStore.kt
@@ -187,32 +187,3 @@ class ReducerException(
  * State update will always be received on [createStore] [CoroutineScope] thread.
  */
 typealias StateReceiver<S> = (newState: S) -> Unit
-
-/**
- * Filters [StateReceiver] by only passing states that are distinct from their predecessor state using [isStateDistinct]
- * function.
- *
- * @param [isStateDistinct] receives `previousState` (can be null initially), `newState` and outputs `true` if new state
- * should be passed to [StateReceiver] or `false` if not.
- */
-fun <S : Any> StateReceiver<S>.distinctUntilChangedBy(
-    isStateDistinct: (previousState: S?, newState: S) -> Boolean
-): StateReceiver<S> {
-    var previousState: S? = null
-
-    return {
-        if (isStateDistinct(previousState, it)) {
-            previousState = it
-            this(it)
-        }
-    }
-}
-
-/**
- * Filters [StateReceiver] by only passing states that are distinct from their predecessor state using states equality,
- * only non-equal states will pass.
- */
-fun <S : Any> StateReceiver<S>.distinctUntilChanged(): StateReceiver<S> =
-    distinctUntilChangedBy { previousState, newState ->
-        previousState != newState
-    }
